catches ctrl-c now... sorta... SEE NOTES \/
This took me WAY longer then it should have. wx trys to handle ctrl-c for us,
but we really don't want that, because It cant do everything for us. I'm going
to have to dig into the wx code to find out how it interacts with sigint, needs
gui interaction to activate the sigint call